[Edit fayure] [Api] change all price to decimal

+++ ECommerce.API/Interface/IHolooArticleRepository.cs
@@ -13,5 +13,5 @@ public interface IHolooArticleRepository : IHolooRepository<HolooArticle>
 
     Task SyncHolooWebId(string aCodeC, int productId, CancellationToken cancellationToken);
     Task<IEnumerable<HolooArticle>> GetHolooArticle(List<string> aCodes, CancellationToken cancellationToken);
-    Task<(int price, double? exist)> GetHolooPrice(string aCode, Price.HolooSellNumber sellPrice);
+    Task<(decimal price, double? exist)> GetHolooPrice(string aCode, Price.HolooSellNumber sellPrice);
 }
